+++ packages/objects/src/primitives/TIckerSymbol.ts
@@ -1,0 +1,4 @@
+import { Brand, make } from "ts-brand";
+
+export type TickerSymbol = Brand<string, "TickerSymbol">;
+export const TickerSymbol = make<TickerSymbol>();
